fix: enable Video playback in AppImage
* set `bundleMediaFramework` for AppImage build in tauri.conf
* install `gstreamer1.0-plugins-bad` and `gstreamer1.0-plugins-ugly` on
  the build host

`bundleMediaFramework` essentially just executes `linuxdeply-plugin-gstreamer`,
which just copies all `gstreamer` files from the build host to the `app.AppDir`
directory, which is then compressed into the AppImage. So we need the
codecs on the build host to get them into the AppImage.

`gstreamer1.0-plugins-bad` is required for H.264 decoding using
`libgstx264.so`, while `gstreamer1.0-plugins-ugly` seems to be required
for `libgsthls.so` and/or `libgstmpegtsdemux.so`.